Fix big repo download
